Add functionality to store an array of random emojis in a session
This commit adds the functionality to store an array of random emojis in a session. The form will be submitted when there are 5 emojis or when the submit button is clicked. Additionally, the commit adds the ability to add multiple contact forms based on the number of emojis clicked.

The commit also includes the following changes:
- Redirect to the index page when logging out
- Allow adding more contacts to the logged-in user

This commit does not include any changes to the graph.